pyproject.toml
[project]
name = "hle-screener"
version = "0.1.0"
description = "HLE-like quality screener for synthetic dataset evaluation"
readme = "README.md"
authors = [
    { name = "caron14", email = "supika.caron@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "datasets>=2.14.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.0",
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "faiss-cpu>=1.7.4",
    "umap-learn>=0.5.0",
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "tqdm>=4.65.0",
    "accelerate>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
